Split the end-to-end tests into 4 test files, each focuses on a seperate test topic.
